e2e: Move helm related functions into helm.go
